requirements.txt
Flask
requests
python-dotenv
pytest
selenium
